Refactor `ColumnCard`
