+++ web-app/client/components/FileStats/Badge/Badge.module.scss
