+++ web-app/client/components/FileStats/ColumnCard/ColumnCard.module.scss
@@ -1,17 +1,9 @@
 @import "styles/common";
 
 .card {
-  background: $white-25;
-  border: 1px solid $black-25;
-  border-radius: 5px;
-  padding: 24px;
   display: flex;
   gap: 24px;
   flex-direction: column;
-
-  @media (max-width: 960px) {
-    padding: 16px;
-  }
 }
 
 .name {
@@ -52,50 +44,13 @@
   gap: 16px;
 }
 
-.stats-blocks {
-  display: flex;
-  gap: 16px 32px;
-  flex-wrap: wrap;
-}
-
-.stats-table {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-
-  th {
-    text-transform: capitalize;
-  }
-
-  tr {
-    display: flex;
-    gap: 32px;
-  }
-
-  td, th {
-    @include paragraph;
-    flex: 1;
-    text-align: left;
-  }
-}
-
-.stats-header {
-  margin-bottom: 8px;
-  color: $primary-0;
-}
-
-.basic-stats {
+.basic-stats, .more-stats {
   display: flex;
   gap: 16px 32px;
   flex-wrap: wrap;
 }
 
 .more-stats {
-  display: flex;
-  gap: 16px 32px;
-  flex-wrap: wrap;
-
   & > * {
     flex: 1;
   }
@@ -104,7 +59,7 @@
     flex-direction: column;
     flex-wrap: initial;
 
-    .stats-blocks {
+    & > article > div {
       justify-content: space-between;
     }
   }
