+++ web-app/client/components/FileStats/Paper/Paper.module.scss
@@ -1,0 +1,12 @@
+@import "styles/common";
+
+.wrapper {
+  background: $white-25;
+  border: 1px solid $black-25;
+  border-radius: 5px;
+  padding: 24px;
+
+  @media (max-width: 960px) {
+    padding: 16px;
+  }
+}
