+++ web-app/client/components/FileStats/Paper/index.ts
@@ -1,0 +1,1 @@
+export * from "./Paper";
