+++ web-app/client/components/FileStats/Statistic/index.ts
@@ -1,0 +1,1 @@
+export * from "./Statistic";
