+++ web-app/client/components/FileStats/StatsBlock/index.ts
@@ -1,0 +1,1 @@
+export * from "./StatsBlock";
